Bug 2049856 - Remove old viaduct backend code
* Removed the old `backend` and `backend::ffi` modules
* Renamed `new_backend` to `backend`.

Changed the default timeout for `viaduct::Client` to 7 seconds for iOS
and 10 seconds otherwise.  This matches the behavior of we had for
`Request::send()`.  Should we make these longer?

Also: removed viaduct-reqwest which was no longer used.  Removed the
`rusqlite::bundled` feature from `ads-client`.  It's not set by other
crates and this fixed a build issue for me.
