+++ Cargo.toml
@@ -53,7 +53,6 @@ members = [
     "components/support/url-macro/tests",
     "components/support/viaduct-dev",
     "components/support/viaduct-hyper",
-    "components/support/viaduct-reqwest",
     "components/sync_manager",
     "components/sync15",
     "components/tabs",
@@ -139,7 +138,6 @@ default-members = [
     "components/support/types",
     "components/support/url-macro",
     "components/support/url-macro/tests",
-    "components/support/viaduct-reqwest",
     "components/sync_manager",
     "components/sync15",
     "components/tabs",
