+++ components/ads-client/Cargo.toml
@@ -16,7 +16,6 @@ error-support = { path = "../support/error" }
 parking_lot = "0.12"
 rusqlite = { version = "0.37.0", features = [
   "functions",
-  "bundled",
   "serde_json",
   "unlock_notify",
 ] }
