+++ components/viaduct/src/client.rs
@@ -3,16 +3,13 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use crate::{
+    backend::get_backend,
     header_names::USER_AGENT,
-    new_backend::get_backend,
     settings::{validate_request, GLOBAL_SETTINGS},
     Request, Response, Result,
 };
 
 /// HTTP Client
-///
-/// This represents the "new" API.
-/// See `README.md` for details about the transition from the old to new API.
 #[derive(Default)]
 pub struct Client {
     settings: ClientSettings,
@@ -106,7 +103,10 @@ impl ClientSettings {
 impl Default for ClientSettings {
     fn default() -> Self {
         Self {
-            timeout: 60000,
+            #[cfg(target_os = "ios")]
+            timeout: 7000,
+            #[cfg(not(target_os = "ios"))]
+            timeout: 10000,
             redirect_limit: 10,
             user_agent: None,
             #[cfg(feature = "ohttp")]
