+++ megazords/ios-rust/focus/Cargo.toml
@@ -12,7 +12,6 @@ crate-type = ["staticlib"]
 rust-log-forwarder = { path = "../../../components/support/rust-log-forwarder" }
 viaduct = { path = "../../../components/viaduct" }
 viaduct-hyper = { path = "../../../components/support/viaduct-hyper" }
-viaduct-reqwest = { path = "../../../components/support/viaduct-reqwest" }
 nimbus-sdk = { path = "../../../components/nimbus" }
 error-support = { path = "../../../components/support/error" }
 remote_settings = { path = "../../../components/remote_settings" }
